requirements.txt
python-dotenv
google-cloud-pubsub
websocket-client